requirements.txt
langchain-community==0.2.5
streamlit==1.35.0
pillow~=10.3.0
python-dotenv==1.0.1